feat: Refactor server structure and implement HTTP handlers for CRUD operations
